+++ src/main/java/eobrazovanje/tim6/app/entity/User.java
@@ -39,7 +39,7 @@ public class User {
 	
 	@Version
 	@Column(name = "version", unique = false, nullable = false)
-	private Integer version;
+	private Long version;
 	
 	@ManyToMany(fetch = FetchType.LAZY)
 	@JoinTable(name = "user_role", joinColumns = @JoinColumn(name = "user_id"), inverseJoinColumns = @JoinColumn(name = "role_id"))
@@ -89,12 +89,12 @@ public class User {
 		deleted = newDeleted;
 	}
 	
-	public Integer getVersion() {
+	public Long getVersion() {
 		
 		return version;
 	}
 	
-	public void setVersion(Integer newVersion) {
+	public void setVersion(Long newVersion) {
 		
 		version = newVersion;
 	}
